ica_module_package/ica_module/requirements.txt
numpy>=1.23
scipy>=1.9
# Optional but strongly recommended for the nuScenes example:
nuscenes-devkit
pyquaternion
pytest
